+++ expensetracker-db-connector/src/main/java/com/slabs/expense/tracker/database/mapper/IncomeMapper.java
@@ -17,7 +17,7 @@ public interface IncomeMapper {
 
 	public List<Income> getIncome(@Param("username") String username, @Param("year") Integer year, @Param("month") Integer month) throws Exception;
 
-	public Float getTotalIncome(@Param("username") String username, @Param("year") Integer year, @Param("month") Integer month) throws Exception;
+	public Double getTotalIncome(@Param("username") String username, @Param("year") Integer year, @Param("month") Integer month) throws Exception;
 
 	public List<IncomeType> getIncomeType() throws Exception;
